Censorship
In lieu of removing the loading tooltip altogether, I've elected to
censor the worst tooltips in the loading screen. It really bothers me
when I get cursed out every time the game loads or run into rape jokes.

While I wish I could remove ALL the curses, that would require me to
find a lot of quotes to replace it. This will be good enough.