Gameplay
Restrict easy stability, war exhaustion and inflation reduction. The means of reduction are now actions in National Policies, and act over time with a gold cost. This is to shift away from the immediate resolution of issues regarding these so the player needs to consider options more carefully when dealing with these attributes.
Added Gambling action that simulates a casino night, allowing the player to gamble.
Added Prostitution action that allows the monarch of a nation to visit a brothel, and perhaps get a heir.