New Vegas
Changed the Mr. House death to fire from the reforms and not use the slave ruler mechanic (no need for it).
Added a ruler flag for Mr. House in history that is used for the death event.
Fixed the incorrect usage of tooltip (instead of custom_tooltip) messing up one of the options.
Added the titles for the New Vegas reforms.
Added the factions into the new_vegas_mechanic reform.