Caesar's Legion
Changed "United Legion" event to ""Caesar, Son of Mars" with three options to pick from.
Added "Alea Iacta Est" event that gives claims on Hoover Dam area or 100 adm as options.
Adapted "The Death of the Caesar" to be more general and for the tribal ruler claim to scale with army tradition