New California Republic
Implemented the HRE as the New California Republic.

The NCR has been split up into the founding states and territories. By default all the territories are flexible subjects of the President, meaning they can war without requiring the overlord to attack.

Every 5 years the NCR has a vote which switches the President between the founding states, which represents the candidate that wins the election. The NCR territory subjects follow the President, meaning losing the presidency means the territories will switch to the new President.

To make the HRE work, the NCR countries are now technically monarchies with heirs. But they all automatically get valid heirs to simulate them selecting candidates.
Due to the hardcoded nature of the HRE, the elector nations are not subjects to the President.

There are three Presidential Orders that can be passed:
- Indefinite Terms: makes the President hereditary
- Centralize Administration: makes all members vassals
- Unify the Republic: forms the NCR tag and inherits all member land